Create new script for updating the environment, properly setting the time
