+++ simulation/utils/__init__.py
@@ -2,3 +2,4 @@ from simulation.utils.InputBounds import InputBounds
 from simulation.utils.SimulationBuilder import SimulationBuilder
 from simulation.utils.Plotting import Plotting, GraphPage
 from simulation.utils.Cryptographer import Key, Encryptor, Decryptor
+from simulation.utils.query import Query
